requirements.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Misaka==1.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
misaka==2.1.1
pycparser==2.20
python-slugify==4.0.1
PyYAML==5.3.1
requests==2.24.0
text-unidecode==1.3
urllib3==1.25.10
Werkzeug==1.0.1
